3.6 | Add category input data validation | Add & update relationship beetwen PostModel and UserModel, CategoryModel

+++ src/app/domain/category/delete_category.py
@@ -1,3 +1,12 @@
+from app.core.exceptions.category_domain_exception import (
+    CategoryNotDeletedException,
+    CategoryNotFoundException,
+    GeneralCategoryCannotBeModifiedException,
+)
+from app.core.exceptions.database_exception import (
+    EntityNotDeletedException,
+    EntityNotFoundException,
+)
 from app.infrastucture.postgresql.database import db
 from app.infrastucture.repositories.category import CategoryRepository
 
@@ -8,5 +17,12 @@ class DeleteCategoryUseCase:
         self._repo = CategoryRepository()
 
     async def execute(self, category_slug: str) -> None:
-        async with self._db.session() as session:
-            await self._repo.delete(session, category_slug)
+        if category_slug == "general":
+            raise GeneralCategoryCannotBeModifiedException()
+        try:
+            async with self._db.session() as session:
+                await self._repo.delete(session, category_slug)
+        except EntityNotFoundException:
+            raise CategoryNotFoundException(category_slug=category_slug)
+        except EntityNotDeletedException:
+            raise CategoryNotDeletedException(category_slug=category_slug)
